requirements.txt
# Groq
groq

# Python dotenv
python-dotenv

# Pydantic
pydantic